Added Optimizations (#9)
* Fixed switching if more than one file pair open.

The previous file is not the associated file if there is more than one buffer open.
For instance: 
    open a.h
    CurtineIncSw() switches to a.c
    open to b.h
    CurtineIncSw() switches to a.c, rather than b.c

Removing this optimization ensures that we always pick the associated file.

* Changing search to be faster by:
  1) Search progressively larger directories
  2) Having find stop after finding one thing

* Reintroduced optimization in a way that didn't break my use case.
Changed optimization to use buffer variables, as opposed to tab variables. From my understanding, tabs are intended to be used as a work space, and can have more than one buffer open. Using buffer variables allows each file to have a link to its associated header/implemenation file.
Removed call to tr, since I'm not sure this is necessary with print.