Fixed corner case when opening other files with find
After switching from a.c to a.h files, opening a file, b.h, with no
corresponding .c file, trying to navigate to the corresponding b.c, and
navigating back to a.h files, CurtineIncSw will rotate between the b.c and a.h.

An example scenario is below:

```
open a.c
CurtineIncSw() switches to a.h
CurtineIncSw() switches to a.c
find b.c (which has no .h)
CurtineIncSw() does nothing
CTRL + O back to a.c
CurtineIncSw() switches to b.c
CurtineIncSw() switches to a.c
etc...
```

This commit fixes that issue. :smile: